# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MutationResult {
  id: string;
  reused?: boolean;
  status?: string;
}

export type CasStateIncidentsItem = { [key: string]: unknown };

export type CasStateActiveIncident = { [key: string]: unknown } | null;

export type CasStateSetupItem = { [key: string]: unknown };

export type CasStateGatesItem = { [key: string]: unknown };

export interface CasState {
  incidents: CasStateIncidentsItem[];
  activeIncident: CasStateActiveIncident;
  setup: CasStateSetupItem[];
  gates: CasStateGatesItem[];
}

export type BootstrapCasReadinessBodyGatesItem = { [key: string]: unknown };

export type BootstrapCasReadinessBodySetupItem = { [key: string]: unknown };

export type BootstrapCasReadinessBody = {
  gates: BootstrapCasReadinessBodyGatesItem[];
  setup: BootstrapCasReadinessBodySetupItem[];
};

export type UpdateCasSetupBody = {
  complete: boolean;
};

export type UpdateCasGateBody = {
  status: string;
};