# lib/api-zod/src/generated/types/casState.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CasStateActiveIncident } from './casStateActiveIncident';
import type { CasStateGatesItem } from './casStateGatesItem';
import type { CasStateIncidentsItem } from './casStateIncidentsItem';
import type { CasStateSetupItem } from './casStateSetupItem';

export interface CasState {
  incidents: CasStateIncidentsItem[];
  activeIncident: CasStateActiveIncident;
  setup: CasStateSetupItem[];
  gates: CasStateGatesItem[];
}